MAILET-115 Move Notify & Redirect mailets in a new package
