+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/DSNBounce.java
@@ -34,6 +34,8 @@ import javax.mail.internet.MimeBodyPart;
 import javax.mail.internet.MimeMessage;
 
 import org.apache.james.core.MailImpl;
+import org.apache.james.transport.mailets.redirect.AbstractRedirect;
+import org.apache.james.transport.mailets.redirect.TypeCode;
 import org.apache.james.transport.util.Patterns;
 import org.apache.mailet.Mail;
 import org.apache.mailet.MailAddress;
