+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/Forward.java
@@ -25,6 +25,8 @@ import javax.mail.MessagingException;
 import javax.mail.internet.AddressException;
 import javax.mail.internet.InternetAddress;
 
+import org.apache.james.transport.mailets.redirect.AbstractRedirect;
+import org.apache.james.transport.mailets.redirect.TypeCode;
 import org.apache.mailet.MailAddress;
 
 import com.google.common.base.Strings;
