+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/NotifyPostmaster.java
@@ -24,6 +24,7 @@ import java.util.Collection;
 import javax.mail.MessagingException;
 import javax.mail.internet.InternetAddress;
 
+import org.apache.james.transport.mailets.redirect.AbstractRedirect;
 import org.apache.mailet.MailAddress;
 import org.apache.mailet.MailetConfig;
